moving numerical tests to the tests directory
