+++ tests/numerical/catastrophic_cancellation.cpp
