+++ tests/numerical/common.hpp
