+++ tests/numerical/contract_expand.cpp
