+++ tests/numerical/exponentiation.cpp
