+++ tests/numerical/paranoia.cpp
