+++ tests/numerical/sterbenz_lemma.cpp
