+++ tests/numerical/sum_of_integers.cpp
