+++ tests/numerical/thin_triangle.cpp
